Initial commit. Some logging helper functions.
The code for colored output and for arg parsing is from some of my other projects.

The code for attaching a log message to an exception is new.
The goal is to log output at critical level if an exception is unhandled.
If the exception is handled, the program can log the buffered log records
at a different level, or discard them.

The default way seems to be logging.handlers.MemoryHandler
However, this approach has one downside: The programmer has to know when
buffering has to be stopped.

My module is useful if you want to attach several lines of debug information to an Exception that you rise
and that may be caught somewhere completely different. If it is not caught, sys.excepthook will log the information.
If it is caught, the code that catches the error may choose to log the information at a different level, e.g. DEBUG